+++ lib/pageResource.ts
